# improve: remove redundant calls of sprintf